# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DeskDibs API",
    "description": "Office hot-desk seat booking. The floor map at GET /api/seatmap is the endpoint the rest of the product hangs off; every mutation broadcasts a SeatStatusChanged message over STOMP to /topic/seatmap/{date} once its transaction commits.",
    "contact": {
      "name": "DeskDibs"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/bookings": {
      "post": {
        "tags": [
          "Bookings"
        ],
        "summary": "Claim a seat for a date",
        "description": "Replaying the same Idempotency-Key returns the original booking instead of a 409.",
        "operationId": "claim",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Client-generated key; a repeat returns the original booking.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimBookingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The seat is now held by the caller.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            }
          },
          "409": {
            "description": "Already booked (by someone else, or by the caller elsewhere).",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            }
          },
          "403": {
            "description": "The seat is held for a team the caller is not in.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            }
          },
          "400": {
            "description": "The date is outside the booking window.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/bookings/{id}/check-in": {
      "post": {
        "tags": [
          "Bookings"
        ],
        "summary": "Check in to today's booking",
        "description": "Only the owner may check in, and only for today's booking.",
        "operationId": "checkIn",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Checked in.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/bookings/move": {
      "post": {
        "tags": [
          "Bookings"
        ],
        "summary": "Move to a different seat on a date",
        "description": "Atomically cancels whatever the caller holds that day and claims the new seat. A target seat that cannot be claimed leaves the original booking untouched and ACTIVE.",
        "operationId": "move",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveBookingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Now holding the new seat.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            }
          },
          "409": {
            "description": "Somebody else won the target seat. The caller keeps the desk they already had.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            }
          },
          "403": {
            "description": "The target seat is held for a team the caller is not in.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            }
          },
          "400": {
            "description": "The date is outside the booking window, or not a working day.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/bookings/me": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "summary": "The caller's own bookings in a date range",
        "description": "Defaults to today through the end of the booking window when from/to are omitted.",
        "operationId": "mine",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "description": "Defaults to today in the office timezone.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "Defaults to the end of the booking window.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The caller's bookings in the range, earliest first.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/bookings/{id}": {
      "delete": {
        "tags": [
          "Bookings"
        ],
        "summary": "Cancel a booking",
        "description": "The owner, the manager of a team the owner belongs to, or an admin may cancel.",
        "operationId": "cancel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The booking is now CANCELLED.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            }
          },
          "403": {
            "description": "The caller does not own this booking.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            }
          },
          "404": {
            "description": "No booking with that id.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ClaimBookingRequest": {
        "type": "object",
        "properties": {
          "seatId": {
            "type": "integer",
            "format": "int64",
            "description": "The seat to claim.",
            "example": 42
          },
          "date": {
            "type": "string",
            "format": "date",
            "description": "ISO date within the booking window.",
            "example": "2026-08-10"
          }
        },
        "required": [
          "date",
          "seatId"
        ]
      },
      "BookingResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "seatId": {
            "type": "integer",
            "format": "int64"
          },
          "seatLabel": {
            "type": "string"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "userDisplayName": {
            "type": "string"
          },
          "bookingDate": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "CANCELLED",
              "RELEASED_NO_SHOW",
              "RELEASED_USER_DEACTIVATED",
              "RELEASED_SEAT_REMOVED"
            ]
          },
          "checkedInAt": {
            "type": "string",
            "format": "date-time"
          },
          "idempotencyKey": {
            "type": "string"
          }
        }
      },
      "MoveBookingRequest": {
        "type": "object",
        "properties": {
          "seatId": {
            "type": "integer",
            "format": "int64",
            "description": "The seat to move to.",
            "example": 43
          },
          "date": {
            "type": "string",
            "format": "date",
            "description": "ISO date within the booking window.",
            "example": "2026-08-10"
          }
        },
        "required": [
          "date",
          "seatId"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Obtain one from POST /api/auth/login (local provider) or the Entra OIDC flow, then send it as \"Authorization: Bearer <token>\".",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}